cleanup after merging
