+++ services/api-server/tests/unit/api_routes_solvers/conftest.py
@@ -3,10 +3,19 @@
 # pylint: disable=unused-variable
 
 
+from typing import Iterator
+
+import _fakes_catalog
 import pytest
+import respx
 from faker import Faker
 from fastapi import FastAPI
 from requests.auth import HTTPBasicAuth
+from respx import MockRouter
+from respx.router import MockRouter
+from simcore_service_api_server.core.settings import ApplicationSettings
+
+from .. import _fakes_catalog
 
 
 @pytest.fixture
@@ -38,3 +47,35 @@ def auth(mocker, app: FastAPI, faker: Faker) -> HTTPBasicAuth:
         return_value=faker.email(),
     )
     return HTTPBasicAuth(faker.word(), faker.password())
+
+
+@pytest.fixture
+def mocked_catalog_service_api(app: FastAPI) -> Iterator[MockRouter]:
+    settings: ApplicationSettings = app.state.settings
+    assert settings.API_SERVER_CATALOG
+
+    # pylint: disable=not-context-manager
+    with respx.mock(
+        base_url=settings.API_SERVER_CATALOG.base_url,
+        assert_all_called=False,
+        assert_all_mocked=True,
+    ) as respx_mock:
+
+        respx_mock.get(
+            "/services?user_id=1&details=false", name="list_services"
+        ).respond(
+            200,
+            json=[
+                # one solver
+                _fakes_catalog.create_service_out(
+                    key="simcore/services/comp/Foo", name="Foo"
+                ),
+                # two version of the same solver
+                _fakes_catalog.create_service_out(version="0.0.1"),
+                _fakes_catalog.create_service_out(version="1.0.1"),
+                # not a solver
+                _fakes_catalog.create_service_out(type="dynamic"),
+            ],
+        )
+
+        yield respx_mock
