+++ services/api-server/tests/unit/api_routes_solvers/test_api_routers_solvers_jobs.py
@@ -2,9 +2,6 @@
 # pylint: disable=unused-argument
 # pylint: disable=unused-variable
 
-from typing import Iterator
-
-
 import urllib
 import urllib.parse
 from pathlib import Path
@@ -18,131 +15,11 @@ import pytest
 import respx
 from faker import Faker
 from fastapi import FastAPI
-from respx import MockRouter
-from simcore_service_api_server.core.settings import ApplicationSettings
 from pydantic import AnyUrl, HttpUrl, parse_obj_as
+from respx import MockRouter
 from respx.router import MockRouter
 from simcore_service_api_server.core.settings import AppSettings
 from starlette import status
-
-
-@pytest.fixture
-def auth(mocker, app: FastAPI, faker: Faker) -> httpx.BasicAuth:
-    # mock engine if db was not init
-    if app.state.settings.API_SERVER_POSTGRES is None:
-
-        engine = mocker.Mock()
-        engine.minsize = 1
-        engine.size = 10
-        engine.freesize = 3
-        engine.maxsize = 10
-        app.state.engine = engine
-
-    # patch authentication entry in repo
-    faker_user_id = faker.pyint()
-
-    mocker.patch(
-        "simcore_service_api_server.db.repositories.api_keys.ApiKeysRepository.get_user_id",
-        return_value=faker_user_id,
-    )
-    mocker.patch(
-        "simcore_service_api_server.db.repositories.users.UsersRepository.get_user_id",
-        return_value=faker_user_id,
-    )
-    mocker.patch(
-        "simcore_service_api_server.db.repositories.users.UsersRepository.get_email_from_user_id",
-        return_value=faker.email(),
-    )
-    return httpx.BasicAuth(faker.word(), faker.password())
-
-
-# test_list_solvers -----------------------------------------------------
-
-
-@pytest.fixture
-def mocked_catalog_service_api(app: FastAPI) -> Iterator[MockRouter]:
-    settings: ApplicationSettings = app.state.settings
-    assert settings.API_SERVER_CATALOG
-
-    # pylint: disable=not-context-manager
-    with respx.mock(
-        base_url=settings.API_SERVER_CATALOG.base_url,
-        assert_all_called=False,
-        assert_all_mocked=True,
-    ) as respx_mock:
-
-        respx_mock.get(
-            "/services?user_id=1&details=false", name="list_services"
-        ).respond(
-            200,
-            json=[
-                # one solver
-                _fakes_catalog.create_service_out(
-                    key="simcore/services/comp/Foo", name="Foo"
-                ),
-                # two version of the same solver
-                _fakes_catalog.create_service_out(version="0.0.1"),
-                _fakes_catalog.create_service_out(version="1.0.1"),
-                # not a solver
-                _fakes_catalog.create_service_out(type="dynamic"),
-            ],
-        )
-
-        yield respx_mock
-
-
-@pytest.mark.skip(reason="Still under development. Currently using fake implementation")
-async def test_list_solvers(
-    client: httpx.AsyncClient,
-    mocked_catalog_service_api: MockRouter,
-    mocker,
-):
-    warn = mocker.patch.object(
-        simcore_service_api_server.api.routes.solvers.logger, "warning"
-    )
-
-    # list solvers latest releases
-    resp = await client.get("/v0/solvers")
-    assert resp.status_code == status.HTTP_200_OK
-
-    # No warnings for ValidationError with the fixture
-    assert (
-        not warn.called
-    ), f"No warnings expected in this fixture, got {str(warn.call_args)}"
-
-    data = resp.json()
-    assert len(data) == 2
-
-    for item in data:
-        solver = Solver(**item)
-        print(solver.json(indent=1, exclude_unset=True))
-
-        # use link to get the same solver
-        assert solver.url.host == "api.testserver.io"  # cli.base_url
-
-        # get_solver_latest_version_by_name
-        resp0 = await client.get(solver.url.path)
-        assert resp0.status_code == status.HTTP_501_NOT_IMPLEMENTED
-        # assert f"GET {solver.name}:{solver.version}"  in resp0.json()["errors"][0]
-        assert f"GET solver {solver.id}" in resp0.json()["errors"][0]
-        # assert Solver(**resp0.json()) == solver
-
-        # get_solver
-        resp1 = await client.get(f"/v0/solvers/{solver.id}")
-        assert resp1.status_code == status.HTTP_501_NOT_IMPLEMENTED
-        assert f"GET solver {solver.id}" in resp1.json()["errors"][0]
-        # assert Solver(**resp1.json()) == solver
-
-        # get_solver_latest_version_by_name
-        resp2 = await client.get(f"/v0/solvers/{solver.id}/latest")
-
-        assert resp2.status_code == status.HTTP_501_NOT_IMPLEMENTED
-        assert f"GET latest {solver.id}" in resp2.json()["errors"][0]
-
-        # assert Solver(**resp2.json()) == Solver(**resp3.json())
-
-
-# test_solver_logs -----------------------------------------------------
 
 
 @pytest.fixture
